Cleanup build tool configuration in all tasks
