+++ jfrog-tasks-utils/package.json
@@ -10,7 +10,8 @@
     "dependencies": {
         "azure-pipelines-task-lib": "^3.1.10",
         "azure-pipelines-tool-lib": "^1.1.0",
-        "typed-rest-client": "^1.8.3"
+        "typed-rest-client": "^1.8.3",
+        "fs-extra": "^7.0.0"
     },
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1"
