+++ jfrog-tasks-utils/utils.d.ts
@@ -35,5 +35,6 @@ declare module '@jfrog/tasks-utils' {
     export function assembleUniqueServerId(usageType: string): string;
     export function configureDefaultXrayServer(usageType: string, cliPath: string, workDir: string): string;
     export function configureDefaultDistributionServer(usageType: string, cliPath: string, workDir: string): string;
+    export function taskDefaultCleanup(cliPath: string, workDir: string, serverIdsArray: string[]): string;
     export { taskSelectedCliVersionEnv };
 }
