+++ tasks/JFrogGradle/gradleBuild.js
@@ -1,7 +1,5 @@
 const tl = require('azure-pipelines-task-lib/task');
 const utils = require('@jfrog/tasks-utils/utils.js');
-const path = require('path');
-const fs = require('fs-extra');
 
 const cliGradleCommand = 'gradle';
 const gradleConfigCommand = 'gradlec';
@@ -103,16 +101,11 @@ function executeGradle(cliPath, workDir) {
 
 function cleanup(cliPath, workDir) {
     // Delete servers.
-    utils.deleteCliServers(cliPath, workDir, [serverIdDeployer, serverIdResolver]);
+    utils.taskDefaultCleanup(cliPath, workDir, [serverIdDeployer, serverIdResolver]);
     // Remove extractor variables.
     try {
         utils.removeExtractorsDownloadVariables(cliPath, workDir);
-        tl.debug('Removing JFrog CLI Gradle configuration');
-        const configPath = path.join(workDir, '.jfrog', 'projects');
-        if (fs.pathExistsSync(configPath)) {
-            fs.removeSync(configPath);
-        }
-    } catch (cleanupException) {
-        tl.setResult(tl.TaskResult.Failed, cleanupException);
+    } catch (removeVariablesException) {
+        tl.setResult(tl.TaskResult.Failed, removeVariablesException);
     }
 }
