+++ tasks/JFrogNuget/nugetBuild.js
@@ -109,9 +109,7 @@ function runNuGet(nugetCommandCli, buildDir, cliPath, configuredServerIdsArray) 
     } catch (ex) {
         tl.setResult(tl.TaskResult.Failed, ex);
     } finally {
-        if (configuredServerIdsArray) {
-            utils.deleteCliServers(cliPath, buildDir, configuredServerIdsArray);
-        }
+        utils.taskDefaultCleanup(cliPath, buildDir, configuredServerIdsArray);
     }
 }
